move program options
